lib: move csc related code to st_convert.c
